+++ ReleaseNotes.md
@@ -2,7 +2,10 @@
 
 ## Release 0.3.8 (NOT RELEASED YET)
 
- * ?
+ * [Fixed Issue 220][issue-220]
+   
+   `getViews()` Do not use `api/json?depth=1` cause timeout.
+
 
 ### API Changes
 
@@ -949,6 +952,7 @@ TestReport testReport = mavenJob.getLastSuccessfulBuild().getTestReport();
 [issue-209]: https://github.com/jenkinsci/java-client-api/issues/209
 [issue-211]: https://github.com/jenkinsci/java-client-api/issues/211
 [issue-215]: https://github.com/jenkinsci/java-client-api/issues/215
+[issue-220]: https://github.com/jenkinsci/java-client-api/issues/220
 [pull-123]: https://github.com/jenkinsci/java-client-api/pull/123
 [pull-149]: https://github.com/jenkinsci/java-client-api/pull/149
 [pull-158]: https://github.com/jenkinsci/java-client-api/pull/158